// components/modal-user/interface.ts
export interface ModalUserLocale {
  selected: string;
  clear: string;
  add: string;
  pleaseholder: string;
  select: string;
  open: string;
  close: string;
  remove: string;
}

export interface ModalUserFieldNames {
  key?: string;
  disabled?: string;
  value?: string;
  children?: string;
  title?: string;
  alias?: string;
  position?: string;
  avatar?: string;
  roleName?: string;
  type?: string;
}